split the scraper cli and the utility code
