+++ web/scrapers/builtin/util/__init__.py
@@ -1,0 +1,8 @@
+from .strings import (
+    remove_special_chars,
+    name_to_id,
+    int_or_none,
+    float_or_none,
+)
+from .scraper import ScraperBase
+from .structs import PoolInfo, LotInfo, LotData
